+++ servo/cli.py
@@ -132,7 +132,7 @@ class Context(typer.Context):
         base_url: Optional[str] = None,
         url: Optional[str] = None,
         **kwargs,
-    ) -> None:
+    ) -> None: # noqa: D107
         self.config_file = config_file
         self.optimizer = optimizer
         self.assembly = assembly
@@ -258,7 +258,7 @@ class CLI(typer.Typer, servo.logging.Mixin):
         callback: Optional[Callable] = typer.models.Default(None),
         section: Section = Section.COMMANDS,
         **kwargs,
-    ):
+    ) -> None: # noqa: D107
 
         # NOTE: Set default command class to get custom context
         if command_type is None:
@@ -594,7 +594,7 @@ class ConnectorCLI(CLI):
         callback: Optional[Callable] = typer.models.Default(None),
         section: Section = Section.COMMANDS,
         **kwargs,
-    ):
+    ) -> None: # noqa: D107
         # Register for automated inclusion in the ServoCLI
         ConnectorCLI.__clis__.add(self)
 
@@ -636,7 +636,7 @@ class ServoCLI(CLI):
         add_completion: bool = True,
         no_args_is_help: bool = True,
         **kwargs,
-    ) -> None:
+    ) -> None: # noqa: D107
         # NOTE: We pass OrderedGroup to suppress sorting of commands alphabetically
         if command_type is None:
             command_type = OrderedGroup
@@ -1832,31 +1832,24 @@ class ServoCLI(CLI):
 
             if short:
                 if format == VersionOutputFormat.text:
-                    typer.echo(
-                        f"{connector_class.full_name} v{connector_class.version}"
-                    )
+                    typer.echo(connector_class.version_summary())
                 elif format == VersionOutputFormat.json:
                     version_info = {
                         "name": connector_class.full_name,
                         "version": str(connector_class.version),
+                        "cryptonym": connector_class.cryptonym,
                     }
                     typer.echo(json.dumps(version_info, indent=2))
                 else:
                     raise typer.BadParameter(f"Unknown format '{format}'")
             else:
                 if format == VersionOutputFormat.text:
-                    typer.echo(
-                        (
-                            f"{connector_class.full_name} v{connector_class.version} ({connector_class.maturity})\n"
-                            f"{connector_class.description}\n"
-                            f"{connector_class.homepage}\n"
-                            f"Licensed under the terms of {connector_class.license}"
-                        )
-                    )
+                    typer.echo(connector_class.summary())
                 elif format == VersionOutputFormat.json:
                     version_info = {
                         "name": connector_class.full_name,
                         "version": str(connector_class.version),
+                        "cryptonym": connector_class.cryptonym,
                         "maturity": str(connector_class.maturity),
                         "description": connector_class.description,
                         "homepage": connector_class.homepage,