+++ servo/connectors/kubernetes.py
@@ -11,7 +11,6 @@ import itertools
 import os
 import pathlib
 import time
-
 from typing import (
     Any,
     Callable,
@@ -36,11 +35,6 @@ from typing import (
 import backoff
 import kubernetes_asyncio
 import pydantic
-
-import kubernetes_asyncio.client
-import kubernetes_asyncio.watch
-import kubernetes_asyncio.client.api_client
-import kubernetes_asyncio.config.kube_config
 
 import servo
 from servo import (
@@ -86,7 +80,7 @@ class Condition(servo.logging.Mixin):
         ValueError: The given ``fn`` is not callable.
     """
 
-    def __init__(self, name: str, fn: Callable, *args, **kwargs) -> None:
+    def __init__(self, name: str, fn: Callable, *args, **kwargs) -> None: # noqa: D107
         if not callable(fn):
             raise ValueError("The Condition function must be callable")
 
@@ -289,7 +283,7 @@ class KubernetesModel(abc.ABC, servo.logging.Mixin):
     is not specified for the resource.
     """
 
-    def __init__(self, obj, **kwargs) -> None:
+    def __init__(self, obj, **kwargs) -> None: # noqa: D107
         self.obj = obj
         self._logger = servo.logger
 
@@ -646,7 +640,7 @@ class Container(servo.logging.Mixin):
         https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.18/#container-v1-core
     """
 
-    def __init__(self, api_object, pod) -> None:
+    def __init__(self, api_object, pod) -> None: # noqa: D107
         self.obj = api_object
         self.pod = pod
 
@@ -2484,7 +2478,12 @@ STANDARD_PERMISSIONS = [
     ),
     PermissionSet(
         group="",
-        resources=["pods", "pods/logs", "pods/status", "namespaces"],
+        resources=["namespaces"],
+        verbs=["get", "list"],
+    ),
+    PermissionSet(
+        group="",
+        resources=["pods", "pods/logs", "pods/status"],
         verbs=["create", "delete", "get", "list", "watch"],
     ),
 ]
@@ -2576,7 +2575,7 @@ class KubernetesConfiguration(BaseKubernetesConfiguration):
             **kwargs,
         )
 
-    def __init__(self, *args, **kwargs) -> None:
+    def __init__(self, *args, **kwargs) -> None: # noqa: D107
         super().__init__(*args, **kwargs)
         self.cascade_common_settings()