+++ servo/entry_points.py
@@ -6,13 +6,18 @@
 # reinstalls of all package dependencies.
 # Do not implement meaningful functionality here. Instead import and
 # dispatch the intent into focused modules to do the real work.
+# noqa
+
 import dotenv
+import uvloop
 
 import servo
 import servo.cli
 
 
 def run_cli() -> None:
+    """Run the Servo CLI."""
+    uvloop.install()
     dotenv.load_dotenv(dotenv.find_dotenv(usecwd=True))
 
     # NOTE: We load connectors here because waiting until assembly