+++ servo/utilities/subprocess.py
@@ -6,7 +6,6 @@ import asyncio
 import datetime
 import pathlib
 import time
-
 from typing import IO, Any, Awaitable, Callable, Dict, List, NamedTuple, Optional, TypeVar, Union, cast
 
 import loguru