+++ tests/test_helpers.py
@@ -6,6 +6,7 @@ from contextlib import contextmanager
 from pathlib import Path
 from typing import Any, Awaitable, Callable, Dict, List, Optional, Type, Union
 
+import uvloop
 import yaml
 from pydantic.json import pydantic_encoder
 
@@ -19,6 +20,7 @@ from servo.servo import Events
 from servo.types import Component, Description, Measurement, RangeSetting
 from servo.utilities import SubprocessResult, Timeout, stream_subprocess_shell
 
+uvloop.install()
 
 class StubBaseConfiguration(BaseConfiguration):
     name: Optional[str]