+++ CHANGELOG.md
@@ -17,6 +17,24 @@ based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). Releases are
 versioned in accordance with [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] "serenity now" - Unreleased
+
+### Added
+
+- Incorporated [uvloop](https://github.com/MagicStack/uvloop) for faster async
+  event loops.
+
+### Changed
+
+- The `__codename__` support has been generalized as `cryptonym` attribute for
+  all connectors.
+- Version output now includes the cryptonym.
+
+### Fixed
+
+- Local versions are now resolved via the `pyproject.toml` file to accurately
+  reflect pre-release versioning and local work in progress.
+
 ## [0.8.3] "pass the calamari" - 2020-10-21
 
 ### Fixed
@@ -25,7 +43,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   that the Kubernetes Connector has an opportunity to set up the environment
   before a measurement is taken that is reliant on the setup (e.g. all canaries
   have been created).
-
+  
 ## [0.8.2] "pass the calamari" - 2020-10-16
 
 ### Added